Schedule missed and new instances onto next 24 hours
The old scheme scheduled checks onto a narrow range of "tomorrow ±2
hours", which created hotspots. Imagine we discover a new instance which
adds 1000 new hosts — all of them will be crammed into that 4-hour
range, and will take a while to decorrelate and spread evenly across the
day. Similarly, if the server is down for a while, then everything that
it didn't check will be scheduled onto just 4 hours, which is a lot of
load.