Let the main thread pick an instance to check
To explain this commit, let us step back for a moment.

The "instances" table contains a list of instances, along with the
timestamp after which the next check should be performed. In essence,
it's a job queue.

Now the question is: how does the program pick the next job to do?

Prior to this commit, the answer was "threads compete to claim the job".
This was implemented as a small SQL transaction where an unclaimed job
is found and then marked as claimed. The "instances" table has
a separate column to mark claimed jobs. This design led to contention on
the writer lock; and since SQLite only has a single, per-database lock,
this also contended with any other writer (e.g. threads that are
currently doing useful work by adding more instances into the database).

To lower the contention, this commit changes the design. The main thread
picks the next instance to check, and puts it into a channel. Worker
threads hold receiving ends of the channel, and compete to receive a job
through it. This competition no longer involves an SQLite lock, so the
contention on the database is reduced. Furthermore, the wait is now much
more efficient, since the waiting threads are sleeping the entire time
(previous design busy-waited with random sleeps).

This also slightly changes the way in which worker threads are added and
removed. In the previous design, if the main tread noticed there's too
much unclaimed work in the DB, it'll spawn a new thread. In the new
design, the channel is has a fixed small-ish capacity, and if the main
thread can't put a new job within a certain amount of time, it spawns
a new thread.

Despite its goal, this new design occasionally suffers from database
lockups >_< I'm still investigating the reasons, but it looks like the
worker threads somehow starve the main thread out of the database.