Fix SQLITE_LOCKED in the main thread
The problem was: after a while, the program would exit claiming that
`db::reschedule()` couldn't update the `instances` table because the
database is locked. The docs for SQLITE_LOCKED[1] explain that this
error comes up either when shared cache is in use, or when an SQL
statement conflicts with something else on the same SQLite connection.
But in our case, neither of this was true: we don't use the shared
cache, and we didn't run any `DROP TABLE`s etc. while running that
`UPDATE`.

We have a main thread and a number of worker threads. Worker threads
hammer at the `instances` all the time, adding new rows and changing
`last_check_datetime` and `state` of the existing ones. The main thread
works at a much lower pace, occasionally running the following:

    BEGIN TRANSACTION;
    SELECT … FROM instances …;
    UPDATE instances SET …;
    COMMIT;

We're running SQLite in WAL mode, which allows readers to co-exist with
writers. (But only one writer is allowed at a time.) Thus, this
transaction may start even if another thread already obtained a writer
lock on the database. From there, a `SELECT` would run just fine: we got
a reader lock, and SQLite provides isolation, so the `SELECT` doesn't
see the effects of that `INSERT`. However, when we get to `UPDATE`,
SQLite tries to upgrade the transaction to a writer lock, and fails. But
not because someone else got the lock! No. That would be SQLITE_BUSY,
and SQLite would handle that for us (by sleeping). No, this causes
SQLITE_LOCKED, because another thread already modified database in such
a way that no matter how long *this* thread waits, and how many times it
re-tries, it won't be able to write its data because it's using a stale
view of the table.

So we just abort, wait 30 seconds, and try again. The main thread is not
in a hurry, it can do this just fine. We try three times, and if we
still get an error, we finally bail.

The fix have been hammering at the database for a couple hours now, and
hasn't failed. I also reproduced this using SQLite CLI tool, and asked
developers to clarify the documentation[3].

1. https://www.sqlite.org/rescode.html#locked
2. https://www.sqlite.org/lang_transaction.html
3. https://sqlite.org/forum/forumpost/bc0ddb7834