Orchestrator: retry on SQLITE_BUSY
This is a continuation to the work that began in
a4faa6abde4d7846cee00a330ab3ee15e69af1f8.

Keith Medcalf told me on SQLite forum[1] that the scenario I described
returns `SQLITE_BUSY_SNAPSHOT`, not `SQLITE_LOCKED`. And this is indeed
the case; I was inattentive, and thought that "database is locked"
message means `SQLITE_LOCKED`, whereas it actually means `SQLITE_BUSY`.

Thus, my analysis in a4faa6abde4d7846cee00a330ab3ee15e69af1f8 was
slightly off (it results in a different error). Here's an update.

As before, the Orchestrator is running the following transaction:

    BEGIN TRANSACTION;
    SELECT … FROM instances …;
    UPDATE instances SET …;
    COMMIT;

At the same time, worker threads are making a lot of `INSERT`s into the
same table. They are doing it outside a transaction, which is to say,
they're doing it inside *implicit* transactions. Since it's `INSERT`
they're doing, the transaction acquires a write lock. We can emulate
this with an explicit transaction:

    BEGIN IMMEDIATE TRANSACTION;
    INSERT INTO instances VALUES …;
    COMMIT;

SQLite returns `SQLITE_BUSY` when `BEGIN IMMEDIATE TRANSACTION` happens
in between `SELECT` and `UPDATE` of the Orcherstrator's transaction.

What happens is:

1. `BEGIN TRANSACTION; SELECT …` obtains a read lock for the
   orchestrator;
2. `BEGIN IMMEDIATE TRANSACTION` obtains a write lock for the worker;
3. `UPDATE` attempts to upgrade orchestrator's lock to a write lock, but
   can't, because the worker is already holding it. So the database
   returns `SQLITE_BUSY`.

Since the error is transient, I chose to ignore it: when it pops up, the
orchestrator goes to sleep for three seconds, and then tries again. The
handler only does this for `SQLITE_BUSY` and `SQLITE_BUSY_*`; all other
errors are immediately propagated upwards, and exit the program.

1. https://sqlite.org/forum/forumpost/bc0ddb7834