Avoid gaps between missed-rescheduled and rescheduled checks
Prior to this commit, we used different periods for:

* rescheduling missed checks (24 hours)
* rescheduling a successful check (29 hours)

As a result, just after a startup we "filled" the next 24 hours with
checks, but as we worked through this backlog, further checks were
scheduled 29 hours from now. As a result, there was a 5-hour gap where
no checks were scheduled, which is bad. (The gap was actually three
hours, because of randomization, but it's a gap nonetheless).

This commit makes the spider use the same 29-hour period in both cases,
closing the gap.