Do not list moving instances if the target is alive
This prevents us from listing "duplicates": an instance and all the
hostnames that redirect to it. As soon as a "main" hostname is marked
"alive", all redirects will be removed from the list. This also removes
any redirect chains (i.e. moving intances that point at instances which
are themselves moving).