Lowercase the domain name before checking it
I saw a domain peer with a domain name '.NET', and it didn't pass the
check because of the case.

It's a good thing to lowercase the domains anyway, to avoid duplicates
(I feel silly because I didn't think of this when I wrote the
architecture description! Such an obvious avenue of attack: mention
a victim host multiple times, but with slightly different
capitalization, and enjoy the DoS attack.)