Fix crash when next_check_datetime is in the past
It's quite normal for the "intended check time" of an instance to be in
the past. A couple situations where this can happen:

- the crawler was restarted, so some checks were not done. Upon startup,
  the crawler notices this and re-schedules them onto the next 24 hours.
  (It could try to make them at once, but there might be a lot of them,
  which would overload the crawler, the server, the network, and the
  checked instances; so we don't do that.)

- if many instances all have the same `next_check_datetime`, the crawler
  could spend more then a second starting the corresponding checks.
  Since the crawler reads one row of the database at a time, it could so
  happen that last rows were read a second after they should've been, so
  their `next_check_datetime` is now in the past.

We used to use chrono, where `a.duration_since(b)` could would return
a negative `Duration` if `a` is older than `b`. We recently migrated to
`std::time` (#293), which returns an `Err` instead. We didn't handle
this new error, so it was propagated and crashed the crawler.

This commit adds handling for this situation. Instead of crashing, the
crawler waits 100ms and fires off the check.

Fixes #330.